Drop half of previous patch, let 'no boolean' delete the node
In most sane setups a YANG boolean is by default false, so a letting a
'no boolean' delete the node, and defaulting to 'false' is a safe bet.

This lets us treat the non-existance of the node as false in case where
we have a must expression checking for 'not(node)', because a node=false
would make the must epxression very unhappy.

The case of a YANG boolean with a default true can still be handled by
'set bolean false', which was always the case.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>