Refactor recent srp_set() changes into srp_set() and srp_helper()
The set command needs to be called synchronous by klish, otherwise the
sysrepo callbacks for update events are not called.

So, we need to factor out the helpers from srp_set() into a "set 2",
which I've dediced to call srp_helper().  In the Infix CLI we call
the helper using two aliases "change" and "editor".

This also fixes an issue where sr_release_context() was not called.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>