srp_commit(): disable implicit "copy running startup"
In the Infix CLI we have adopted the Cisco style explicit call to save
running-config to persistent storage with a dedicated copy command in
top-level (admin-exec) context:

    jacky@example:/> copy running-config startup-config

Beneficial for cases when activating the candidate changes to running
lead to loss of the device.  A simple power-cycle and the user is back
running using a working startup-config.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>